refactor: split setup steps and create migrations to move test settings and legacy test defaults
