+++ components/ILIAS/Test/classes/class.ilTestAccess.php
@@ -20,6 +20,7 @@ declare(strict_types=1);
 
 use ILIAS\Test\Participants\ParticipantRepository;
 use ILIAS\Test\Participants\Participant;
+use ILIAS\Test\Settings\SettingsNotFoundException;
 use ILIAS\Test\TestDIC;
 use ILIAS\Test\Access\ParticipantAccess;
 use ILIAS\Test\Settings\MainSettings\MainSettingsDatabaseRepository;
@@ -187,6 +188,8 @@ class ilTestAccess
         try {
             $access_settings = $this->main_settings_repository->getForObjFi($obj_id)
                 ->getAccessSettings();
+        } catch (SettingsNotFoundException $e) {
+            return ParticipantAccess::MISSING_SETTINGS;
         } catch (\Exception $e) {
             return ParticipantAccess::BROKEN_TEST;
         }
