+++ components/ILIAS/Test/src/Access/ParticipantAccess.php
@@ -29,6 +29,7 @@ enum ParticipantAccess: string
     case INDIVIDUAL_CLIENT_IP_MISMATCH = 'individual_client_ip_mismatch';
     case TEST_LEVEL_CLIENT_IP_MISMATCH = 'test_level_client_ip_mismatch';
     case BROKEN_TEST = 'object_is_broken';
+    case MISSING_SETTINGS = 'settings_not_found';
 
     public function getAccessForbiddenMessage(Language $lng): string
     {
@@ -37,6 +38,7 @@ enum ParticipantAccess: string
             self::INDIVIDUAL_CLIENT_IP_MISMATCH,
             self::TEST_LEVEL_CLIENT_IP_MISMATCH => $lng->txt('user_ip_outside_range'),
             self::BROKEN_TEST => $lng->txt('broken_test'),
+            self::MISSING_SETTINGS => $lng->txt('tst_settings_not_found'),
             default => ''
         };
     }
